Fix: Float error

+++ epidec/datasets/swuniv_dacon/__init__.py
@@ -118,7 +118,7 @@ class BalancedSWUnivDaconDataset(SWUnivDaconDataset):
 
         if self.is_train:
             ai_count = len(raw[raw['generated'] == 1])
-            humans = raw[raw['generated'] == 0].sample(n=ai_count * self.balancing_ratio, random_state=seed)
+            humans = raw[raw['generated'] == 0].sample(n=int(ai_count * self.balancing_ratio), random_state=seed)
             raw = pd.concat([humans, raw[raw['generated'] == 1]], ignore_index=False)
             return raw['full_text'].tolist(), raw['generated'].tolist(), raw
         else:
